Refactored mafia bot. New class for vote count
